Fixing multer middleware and user controllers

+++ controller/userControler.js
@@ -5,10 +5,13 @@ const { UserModel } = db
 
 export const createUser = async (req, res) => {
     const { firstName, lastName, dob, email, password, role } = req.body
-    const image = req.file?.path
+    const image = req.file.path
 
     if (!firstName || !lastName || !dob || !email || !password || !role)
         res.status(400).send('All fields are required!')
+    if(!req.file) {
+        return res.status(400).json({error : "Please upload an image"})
+    }
     try {
         const newUser = await UserModel.create({
             firstName,
@@ -32,19 +35,18 @@ export const showAllUsers = async (req, res) => {
     const offset = (page - 1) * pageSize
     try {
         const users = await UserModel.findAll({
-            include: [UserModel],
             offset,
             limit: parseInt(pageSize)
         })
         res.status(200).json({ Users: users })
     }
     catch (error) {
-        res.status(500).send(error)
+        res.status(500).json({error : error})
     }
 }
 
 export const showOneUser = async (req, res) => {
-    const id = req.body.id
+    const id = req.params.id
     try {
         const user = await UserModel.findOne({ where: { id: id } })
         if (user)
@@ -58,35 +60,46 @@ export const showOneUser = async (req, res) => {
 }
 
 export const updateUser = async (req, res) => {
-    const id = req.body.id
-    const { firstName, lastName, dob, email, password, role } = req.body
-    const newImage = req.file?.path
-    if (!firstName || !lastName || !dob || !email || !password || !role)
-        res.status(400).send('All fields are required!')
+    const id = req.body.id;
+    const { firstName, lastName, dob, email, password, role } = req.body;
+    const newImage = req.file?.path;
+
+    if (!firstName || !lastName || !dob || !email || !password || !role) {
+        res.status(400).send('All fields are required!');
+        return;
+    }
+
     try {
-        const user = await UserModel.findOne({ where: { id: id } })
-        const oldImage = user.image
-        if (!user)
-            res.status(404).send(`User ${id} does not exist!`)
-        const editUser = await UserModel.update({
-            firstName,
-            lastName,
-            dob,
-            email,
-            password,
-            newImage,
-            role
-        })
-        await fs.unlink(oldImage)
-        if (editUser)
-            res.status(200).send(`User ${id} has been updated successfully!`)
-        else
-            res.status(404).send(`User ${id} does not exist!`)
+        const user = await UserModel.findOne({ where: { id: id } });
+        const oldImage = user.image;
+
+        if (!user) {
+            res.status(404).send(`User ${id} does not exist!`);
+            return;
+        }
+
+        const editUser = await UserModel.update(
+            {
+                firstName,
+                lastName,
+                dob,
+                email,
+                password,
+                image: newImage,
+                role
+            },
+            {
+                where: { id: id }
+            }
+        ); 
+        if (req.file) {
+            await fs.unlink(oldImage);
+        }
+        res.status(200).send(`User ${user.firstName} has been updated successfully!`);
+    } catch (error) {
+        res.status(500).send(error);
     }
-    catch (error) {
-        res.status(500).send(error)
-    }
-}
+};
 
 export const deleteUser = async (req, res) => {
     const id = req.body.id
@@ -95,9 +108,9 @@ export const deleteUser = async (req, res) => {
         if (!user)
             res.status(404).send(`User ${id} does not exist!`)
         await user.destroy()
-        res.status(200).send(`User ${id} has been deleted successfully!`)
+        res.status(200).json({message : `User ${id} has been deleted successfully!`})
     }
     catch (error) {
-        res.status(500).send(error)
+        res.status(500).json({error: error})
     }
 }
